feat: add `--puppeteer-ignore-default-args` option
